adds modular coordinate system, implements on player and treasure
In order to handle the looping/wrapping map system, this commit implements a global static function that takes an "internal position", which is where the node would be if it had continued traveling in that direction, and returning an "external position", which is where the node is actually placed in the scene.

In this proposal, all nodes that loop/wrap around the map should have an internal_pos Vector2, and all movement code for said nodes should be performed on the internal_pos, and then used to set the external position:
`{node}.position = ModCoord.get_external_pos({node}.internal_pos)`

This commit also removes the old wrapping code, and changes the spawn order of players so they have a position assigned before the enter the tree.